requirements.txt
ipdb==0.13.13
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0
scikit_learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
submitit==1.5.1
timm==0.3.2
torch==1.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.66.2
